[refactor] #25 - 내정보 화면 뷰 파일 분리
